requirements.txt
pillow
opencv-python
tensorflow
numpy
ipython
twilio
tensorflow-object-detection-api
six
